+++ Business/BasePage.cs
@@ -1,10 +1,4 @@
 ﻿using OpenQA.Selenium;
-using SeleniumExtras.PageObjects;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using TAF_ReportPortal_Configuration;
 
 namespace TAF_ReportPortal_Business
