+++ Business/Menu.cs
@@ -1,11 +1,5 @@
 ﻿using OpenQA.Selenium;
 using SeleniumExtras.PageObjects;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using TAF_ReportPortal_Configuration;
 
 namespace TAF_ReportPortal_Business
 {
